+++ Core/Libraries/Source/WWVegas/WWDownload/registry.cpp
@@ -28,8 +28,13 @@ bool  getStringFromRegistry(HKEY root, std::string path, std::string key, std::s
 {
 	HKEY handle;
 	unsigned char buffer[256];
-	unsigned long size = 256;
-	unsigned long type;
+	// TheSuperHackers @fix macOS-port-LP64-sweep: Win32 RegQueryValueEx takes
+	// `LPDWORD` (DWORD* = uint32_t* per the typedef sweep). `unsigned long` is
+	// 8 bytes on macOS LP64 — was silently accepted before only because DWORD
+	// itself was `unsigned long`. Now DWORD = uint32_t, so the call requires
+	// matching types here. No-op on Windows where unsigned long is 4 bytes.
+	DWORD size = 256;
+	DWORD type;
 	int returnValue;
 
 	if ((returnValue = RegOpenKeyEx( root, path.c_str(), 0, KEY_READ, &handle )) == ERROR_SUCCESS)
@@ -50,9 +55,10 @@ bool  getStringFromRegistry(HKEY root, std::string path, std::string key, std::s
 bool getUnsignedIntFromRegistry(HKEY root, std::string path, std::string key, unsigned int& val)
 {
 	HKEY handle;
-	unsigned long buffer;
-	unsigned long size = sizeof(buffer);
-	unsigned long type;
+	// LP64 sweep: see note above.
+	DWORD buffer;
+	DWORD size = sizeof(buffer);
+	DWORD type;
 	int returnValue;
 
 	if ((returnValue = RegOpenKeyEx( root, path.c_str(), 0, KEY_READ, &handle )) == ERROR_SUCCESS)
